Avoid circular import in py2
